+++ lib/axn/core/contract_for_subfields.rb
@@ -67,11 +67,12 @@ module Axn
           allow_blank ||= optional
 
           _parse_field_validations(*fields, allow_nil:, allow_blank:, **validations).map do |field, parsed_validations|
-            if readers
-              _define_subfield_reader(field, on:, validations: parsed_validations)
-              _define_boolean_predicate_reader(field) if _boolean_field?(parsed_validations)
+            SubfieldConfig.new(field:, validations: parsed_validations, on:, sensitive:, preprocess:, default:, metadata:).tap do |config|
+              if readers
+                _define_subfield_reader(field, on:, validations: parsed_validations)
+                _define_boolean_predicate_reader(field) if Axn::Internal::FieldConfig.boolean?(config)
+              end
             end
-            SubfieldConfig.new(field:, validations: parsed_validations, on:, sensitive:, preprocess:, default:, metadata:)
           end
         end